Move tours schema to schema directory (to fix package structure)
Discovered in https://github.com/galaxyproject/galaxy/pull/20868#issuecomment-3338402529
