+++ lib/galaxy/schema/tours.py
