+++ lib/galaxy/tours/__init__.py
@@ -1,11 +1,4 @@
 from ._impl import build_tours_registry
 from ._interface import ToursRegistry
-from ._schema import (
-    Tour,
-    TourCore,
-    TourDetails,
-    TourList,
-    TourStep,
-)
 
-__all__ = ["build_tours_registry", "ToursRegistry", "Tour", "TourCore", "TourDetails", "TourList", "TourStep"]
+__all__ = ["build_tours_registry", "ToursRegistry"]
